+++ javadoc/datagenerator/iaqsensor/TimeUnit.html
@@ -127,7 +127,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Enum&lt;<a href="TimeUnit.html" title="enum in datagenerator.rangedsensor.iaqsensor">TimeUnit</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>datagenerator.rangedsensor.api.TimeUnit</li>
+<li>util.time.TimeUnit</li>
 </ul>
 </li>
 </ul>
